move login components to components dir
