+++ src/components/authentication/types.ts
@@ -1,0 +1,1 @@
+export type LoginStage = "userCheck" | "connectWallet" | "verify" | "signIn";
