+++ src/components/index.ts
@@ -14,3 +14,4 @@ export * from "./bullet-point/index.js";
 export * from "./player-profiles/index.js";
 export * from "./points-and-penalties/index.js";
 export * from "./game-rules/index.js";
+export * from "./authentication/index.js";
